Set test path for realm; finish fixing tests
